drivers/misc/qemu-edu: Fix qemu edu driver llvm build failure
pin_user_pages_fast expects an unsigned long, not a pointer, simply cast
to fix the following build failure:

../drivers/misc/qemu-edu.c:97:29: error: incompatible pointer to integer conversion passing 'const char *' to parameter of type 'unsigned long' [-Wint-conversion]
                ret = pin_user_pages_fast(buf, 1, FOLL_WRITE, pages);
                                          ^~~
../include/linux/mm.h:2384:39: note: passing argument to parameter 'start' here
int pin_user_pages_fast(unsigned long start, int nr_pages,
                                      ^
../drivers/misc/qemu-edu.c:149:29: error: incompatible pointer to integer conversion passing 'char *' to parameter of type 'unsigned long' [-Wint-conversion]
                ret = pin_user_pages_fast(buf, 1, 0, pages);
                                          ^~~
../include/linux/mm.h:2384:39: note: passing argument to parameter 'start' here
int pin_user_pages_fast(unsigned long start, int nr_pages,
                                      ^

+++ drivers/misc/qemu-edu.c
@@ -94,7 +94,7 @@ static ssize_t qemu_edu_write(struct file *fp, const char __user *buf, size_t co
 		return -EINVAL;
 
 	if (!ctx->sva) {
-		ret = pin_user_pages_fast(buf, 1, FOLL_WRITE, pages);
+		ret = pin_user_pages_fast((unsigned long)buf, 1, FOLL_WRITE, pages);
 		if (ret != 1) {
 			pr_err("Failure locking pages.\n");
 			return -ENOMEM;
@@ -146,7 +146,7 @@ static ssize_t qemu_edu_read(struct file *fp, char __user *buf, size_t count, lo
 		return -EINVAL;
 
 	if (!ctx->sva) {
-		ret = pin_user_pages_fast(buf, 1, 0, pages);
+		ret = pin_user_pages_fast((unsigned long)buf, 1, 0, pages);
 		if (ret != 1) {
 			pr_err("Failure locking pages.\n");
 			return -ENOMEM;
